Exp_struct: Add tag field
The tag field can be used as a general expression tag.
In this branch only dialect component tag will be implemented.

To keep 32-byte size on 64-bit machines, change the cost from double to
float, and the type field to 8-bits.

+++ link-grammar/dict-file/read-dict.c
@@ -1187,7 +1187,7 @@ static Exp *make_expression(Dictionary dict)
 			 */
 			if (is_number(dict->token))
 			{
-				double cost;
+				float cost;
 
 				if (strtodC(dict->token, &cost))
 				{
