+++ link-grammar/utilities.h
@@ -504,7 +504,7 @@ size_t get_max_space_used(void);
 char * get_default_locale(void);
 void set_utf8_program_locale(void);
 bool try_locale(const char *);
-bool strtodC(const char *, double *);
+bool strtodC(const char *, float *);
 
 /**
  * Returns the smallest power of two that is at least i and at least 1
